2461-maximum-sum-of-distinct-subarrays-with-length-k 08-23-2025

+++ LeetCode/Medium/2461-maximum-sum-of-distinct-subarrays-with-length-k/2461-maximum-sum-of-distinct-subarrays-with-length-k.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     long long maximumSubarraySum(vector<int>& nums, int k) {
-        int maxSum = 0;
+        long long maxSum = 0;
         unordered_set<int> seen;
         int left = 0;
         int right = 0;
